Refac: Add ImageEncodable Module for Avatar and Quiz
